Update public/recipe_comments, recipe_materials, recipe_steps 他ユーザーの投稿削除抑止機能追加

+++ app/controllers/public/recipe_materials_controller.rb
@@ -35,8 +35,12 @@ class Public::RecipeMaterialsController < ApplicationController
   def destroy
     recipe = Recipe.find(params[:recipe_id])
     recipe_material = RecipeMaterial.find(params[:id])
-    recipe_material.destroy
-    redirect_to new_recipe_recipe_material_path(recipe)
+    if recipe.user != current_user
+      redirect_to recipe_path(recipe)
+    else
+      recipe_material.destroy
+      redirect_to new_recipe_recipe_material_path(recipe)
+    end
   end
   
   private
